Update model for Hexagonal.
