+++ src/model/interfaces/client-i.model.ts
@@ -1,7 +1,7 @@
 import {Observable} from 'rxjs';
-import {JobPostModel} from '../job-post.card';
 import {RecommendationModel} from '../recommendation.model';
 import {UserModel} from '../user.model';
+import {JobPostModel} from '../job-post.model';
 
 export interface ClientI {
   getRecommendations(user: UserModel): Observable<RecommendationModel>;
