+++ src/pages/home/entity/job-post.card.ts
@@ -1,0 +1,6 @@
+export class JobPostCard {
+    uuid?: string;
+    image: string;
+    title: string;
+    shortDescription: string;
+}
