+++ src/utils/job.utils.ts
@@ -1,11 +1,12 @@
-import {JobPostCard, JobPostModel} from '../model/job-post.card';
 import {RecommendationModel} from '../model/recommendation.model';
+import {JobPostCard} from '../pages/home/entity/job-post.card';
+import {JobPostModel} from '../model/job-post.model';
 
 export class JobUtils {
-  static mapJobPostCardToJobPostModel(
-    recommendations: RecommendationModel,
-    card: JobPostCard
-  ): JobPostModel {
-    return recommendations.recommendedPosts.find(rec => rec.uuid === card.uuid);
-  }
+    static mapJobPostCardToJobPostModel(
+        recommendations: RecommendationModel,
+        card: JobPostCard
+    ): JobPostModel {
+        return recommendations.recommendedPosts.find(rec => rec.uuid === card.uuid);
+    }
 }
